# syncro/apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Syncro API",
    "description": "Syncro backend REST API.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/machines/{machineId}": {
      "get": {
        "tags": [
          "machine-controller"
        ],
        "summary": "Get machine",
        "operationId": "getMachine",
        "parameters": [
          {
            "name": "machineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Machine returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid machine id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "404": {
            "description": "Machine not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "machine-controller"
        ],
        "summary": "Update machine",
        "operationId": "updateMachine",
        "parameters": [
          {
            "name": "machineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Machine updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "400": {
            "description": "Validation, malformed JSON, duplicate code, invalid status, or invalid relation",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "404": {
            "description": "Machine, plant, or machine group not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          },
          "409": {
            "description": "Machine data integrity conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "machine-controller"
        ],
        "summary": "Delete machine",
        "operationId": "deleteMachine",
        "parameters": [
          {
            "name": "machineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Machine deleted"
          },
          "400": {
            "description": "Invalid machine id"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Machine not found"
          },
          "409": {
            "description": "Machine data integrity conflict"
          }
        }
      }
    },
...
    "/api/v1/machine-groups/{machineGroupId}": {
      "get": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "Get machine group",
        "operationId": "getMachineGroup",
        "parameters": [
          {
            "name": "machineGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Machine group returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid machine group id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "404": {
            "description": "Machine group not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "Update machine group",
        "operationId": "updateMachineGroup",
        "parameters": [
          {
            "name": "machineGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Machine group updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "400": {
            "description": "Validation, malformed JSON, duplicate machine group name, or invalid machine group id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "404": {
            "description": "Machine group or plant not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "409": {
            "description": "Machine group data integrity conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "Delete machine group",
        "operationId": "deleteMachineGroup",
        "parameters": [
          {
            "name": "machineGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Machine group deleted"
          },
          "400": {
            "description": "Invalid machine group id"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Machine group not found"
          },
          "409": {
            "description": "Machine group data integrity conflict"
          }
        }
      }
    },
...
    "/api/v1/machine-groups": {
      "get": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "List machine groups",
        "operationId": "listMachineGroups",
        "parameters": [
          {
            "name": "plantId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name,asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Machine groups returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid plant id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupListResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Plant not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "Create machine group",
        "operationId": "createMachineGroup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Machine group created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "400": {
            "description": "Validation, malformed JSON, duplicate machine group name, or invalid plant id",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "404": {
            "description": "Plant not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          },
          "409": {
            "description": "Machine group data integrity conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MachineGroupView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/machine-groups/{machineGroupId}/section": {
      "put": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "Assign the machine group to a section (reassignment rejected)",
        "operationId": "assignMachineGroupSection",
        "parameters": [
          {
            "name": "machineGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Assigned"
          }
        }
      },
      "delete": {
        "tags": [
          "machine-group-controller"
        ],
        "summary": "Clear the machine group's section assignment",
        "operationId": "clearMachineGroupSection",
        "parameters": [
          {
            "name": "machineGroupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cleared"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MachineRequest": {
        "type": "object",
        "properties": {
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "machineGroupId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 0,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "brand": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "installedAt": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "optionalTelemetryFields": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 0
          }
        },
        "required": [
          "code",
          "machineGroupId",
          "plantId",
          "status"
        ]
      },
      "MachineView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "plantCode": {
            "type": "string"
          },
          "plantName": {
            "type": "string"
          },
          "machineGroupId": {
            "type": "string",
            "format": "uuid"
          },
          "machineGroupName": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "brand": {
            "type": "string"
          },
          "installedAt": {
            "type": "string",
            "format": "date"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "optionalTelemetryFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "latestTelemetry": {
            "$ref": "#/components/schemas/TelemetryData"
          }
        }
      },
      "TelemetryData": {
        "type": "object",
        "properties": {
          "machineId": {
            "type": "string",
            "format": "uuid"
          },
          "running": {
            "type": "boolean"
          },
          "runtimeHours": {
            "type": "number",
            "format": "double"
          },
          "counting": {
            "type": "integer",
            "format": "int64"
          },
          "lastReceivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "freshnessState": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "STALE"
            ]
          },
          "optionalFields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "hasOptionalFields": {
            "type": "boolean"
          }
        }
      },
...
      "MachineGroupRequest": {
        "type": "object",
        "properties": {
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "name",
          "plantId"
        ]
      },
      "MachineGroupView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "plantCode": {
            "type": "string"
          },
          "plantName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "sectionCode": {
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          }
        }
      },
...
      "MachineGroupListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MachineGroupView"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "string"
          }
        }
      },
...
      "AssignSectionRequest": {
        "type": "object",
        "required": [
          "sectionId"
        ],
        "properties": {
          "sectionId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}